+++ src/lib/constants.ts
@@ -1,0 +1,10 @@
+export const DEFAULT_RESPONSE_CODE = 200;
+
+export const DEFAULT_RESPONSE_HEADERS = {
+	"Content-Type": "text/event-stream",
+	"Cache-Control":
+		"private, no-cache, no-store, no-transform, must-revalidate, max-age=0",
+	Connection: "keep-alive",
+	Pragma: "no-cache",
+	"X-Accel-Buffering": "no",
+};